requirements.txt
pynput==1.7.7
pywin32==308
PyYAML==5.4.1
PyYAML==6.0.2
